Dataframe class is working
