+++ .gitignore
@@ -1,0 +1,2 @@
+test.py
+__dataframe_file.xlsx
