+++ data_structures.py
@@ -1,11 +1,15 @@
+from datetime import datetime
+import pandas as pd
+from parsers import ParserDataFrame, ParserGoogleSheet
+
 class DataFrame:
     def __init__(self, resouce, tables=[]):
         self.__tables = tables
-        self.__parse_global = ParserGoogleSheet(resouce, "__dataframe_file.xlsx")
+        self.__parser_global = ParserGoogleSheet(resouce, "__dataframe_file.xlsx")
 
         self.__update()
 
-        self.__parser_local = Parser__dataFrame(self.__tables)
+        self.__parser_local = ParserDataFrame(self.__tables)
         self.__last_update = datetime.now()
 
     def add(self, table):
@@ -39,41 +43,3 @@ class DataFrame:
             self.__update()
         return True
 
-
-class Table:
-    def __init__(self, name, columns, lines=[], is_values_int=False):
-        self.__name = name
-        self.__columns = columns
-        if not lines:
-            lines = [pd.Series()] * len(columns)
-        self.__data = pd.DataFrame(self.__make_dir(columns, lines))
-        if is_values_int:
-            self.__data = self.__data.convert_dtypes()
-
-    def add_line(self, line):
-        self.__data = self.__data.append(self.__make_dir(self.__data.columns, line), ignore_index=True)
-
-    def add_lines(self, columns):
-        for line in zip(*columns):
-            self.add_line(line)
-
-    def get_data(self):
-        return self.__data
-
-    def get_name(self):
-        return self.__name
-
-    def sort_by(self, param):
-        self.__data = self.__data.sort_values(by=param)
-
-    def reset_index(self):
-        self.__data = self.__data.reset_index()
-        del self.__data["index"]
-
-    def __make_dir(self, keys, values):
-        return {k: v for k, v in zip(keys, values)}
-
-    def __getitem__(self, item):
-        for c in self.__colomns:
-            if item in list(self.__data[c]):
-                return self.__data[self.__data[c] == item]
