+++ global_parser_utilites.py
@@ -1,4 +1,5 @@
 import pandas as pd
+from table import Table
 
 
 def expand_line(n, symbol):
@@ -8,7 +9,7 @@ def expand_line(n, symbol):
 
 class GlobalParserUtility:
     def __init__(self, global_parser):
-        self.__global_parser = global_parser
+        self._global_parser = global_parser
 
     def parse(self):
         pass
@@ -19,29 +20,29 @@ class TeacherTablePU(GlobalParserUtility):
         super().__init__(global_parser)
 
     def parse(self):
-        try:
-            sheet_values = pd.concat([
-                self.__get_teacher_and_post("Список преподавателей институск"),
-                self.__get_teacher_and_post("Список преподавателей баз.кафед")])
-            teachers = Table("teachers", ['surname', 'name', 'lastname', 'post'])
+        #try:
+        sheet_values = pd.concat([
+            self.__get_teacher_and_post("Список преподавателей институск"),
+            self.__get_teacher_and_post("Список преподавателей баз.кафед")])
+        teachers = Table("teachers", ['surname', 'name', 'lastname', 'post'])
 
-            for i in range(len(sheet_values)):
-                try:
-                    teachers.add_line(
-                        sheet_values.iloc[i]["Преподаватель"].split() + [sheet_values.iloc[i]["Должность"].lower()])
-                except Exception as e:
-                    print(sheet_values.iloc[i])
+        for i in range(len(sheet_values)):
+            try:
+                teachers.add_line(
+                    sheet_values.iloc[i]["Преподаватель"].split() + [sheet_values.iloc[i]["Должность"].lower()])
+            except Exception as e:
+                print(sheet_values.iloc[i])
 
-            teachers.sort_by("surname")
-            teachers.reset_index()
-            print("ParserTeacherTable: completed")
-            return teachers
-        except Exception as e:
-            print("ParserTeacherTable: WARNING!")
-            print(e)
+        teachers.sort_by("surname")
+        teachers.reset_index()
+        print("ParserTeacherTable: completed")
+        return teachers
+        #except Exception as e:
+         #   print("ParserTeacherTable: WARNING!")
+          #  print(e)
 
     def __get_teacher_and_post(self, sheet_name):
-        df = self.__global_parser.get_sheet(sheet_name)
+        df = self._global_parser.get_sheet(sheet_name)
         df["Преподаватель"] = df["Преподаватель"].apply(expand_line(3, "-"))
         df["Должность"] = df["Должность"].apply(expand_line(1, "-"))
         teachers = df[["Преподаватель", "Должность"]]
@@ -54,7 +55,7 @@ class ClassTablePU(GlobalParserUtility):
 
     def parse(self):
         try:
-            sheet_values = self.__global_parser.get_sheet("аудиторный фонд")
+            sheet_values = self._global_parser.get_sheet("аудиторный фонд")
             classes = list(filter(lambda x: str(x).isdigit(), sheet_values.loc[0]))
             print("ParserClassTable: completed")
             return Table("classes", ["number"], [classes])
@@ -71,7 +72,7 @@ class GroupTablePU(GlobalParserUtility):
         try:
             groups = Table("groups", ['course', 'group', 'profile'])
             for num in range(1, 7):
-                sheet_values = self.__global_parser.get_sheet(str(num) + " курс ")
+                sheet_values = self._global_parser.get_sheet(str(num) + " курс ")
                 group_names = list(filter(lambda x: x and not str(x).isalpha(), sheet_values.loc[1]))
                 groups.add_lines(
                     [[num] * len(group_names), group_names, sheet_values.loc[2][2:]])  # (course, group_name)
