+++ main.py
@@ -1,10 +1,3 @@
-import numpy as np
-import pandas as pd
-import json
-from datetime import datetime
-
-import httplib2
-from oauth2client.service_account import ServiceAccountCredentials
-
-import math
-import pandas as pd
+from data_structures import DataFrame
+tf = DataFrame("1s_u2pPZ3xdu_tBrVy7hriV2xj15OP9evJfVAuzFyZSc") #1s_u2pPZ3xdu_tBrVy7hriV2xj15OP9evJfVAuzFyZSc")
+tf.get_tables()
